requirements.txt
psycopg2-binary 
pandas
BeautifulSoup4
requests
Django==3.2.10
djangorestframework 
lxml
APScheduler
selenium